requirements.txt

azul-runner>=4.0.0
pillow